+++ package.json
@@ -21,11 +21,13 @@
     "@react-navigation/native-stack": "^7.1.9",
     "@reduxjs/toolkit": "^2.8.2",
     "formik": "^2.4.6",
-    "i18next": "^25.3.2",
+    "i": "^0.3.7",
+    "i18next": "^24.2.2",
     "i18next-browser-languagedetector": "^8.2.0",
     "i18next-http-backend": "^3.0.2",
+    "npm": "^11.5.1",
     "react": "18.3.1",
-    "react-i18next": "^15.6.0",
+    "react-i18next": "^15.1.1",
     "react-native": "0.76.7",
     "react-native-fbsdk-next": "^13.4.1",
     "react-native-paper": "^5.14.5",
