+++ src/navigations/App/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import Tabs from './TabNavigator';
-import { HomeStackParamList } from '@types/navigations';
+import { HomeStackParamList } from '../../types/navigations';
 
 const Stack = createNativeStackNavigator<HomeStackParamList>();
