+++ src/navigations/index.tsx
@@ -4,11 +4,11 @@ import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from '../redux/store';
 import i18n from '../assets/i18n';
-import SplashScreen from '@screens/Auth/Splash';
 import { setDarkMode } from '@redux/slices/ThemeSlice';
-import HomeNavigator from './App';
 import AuthNavigator from './Auth';
+import HomeNavigator from './App';
 import colors from '@assets/theme';
+import Splash from '@screens/Auth/Splash';
 
 const lightTheme = {
   ...DefaultTheme,
@@ -56,7 +56,14 @@ const StackNavigator: React.FC = () => {
     [isDarkMode],
   );
 
-  if (isSplashVisible) return <SplashScreen />;
+  if (isSplashVisible) {
+    return (
+      <NavigationContainer>
+        <Splash />
+      </NavigationContainer>
+    );
+  }
+ 
 
   return (
     <NavigationContainer theme={appTheme}>
